♻️ refactor : 폴더 재구조화 - dto 내부 도메인별 폴더 생성
