+++ src/main/java/com/zip/community/platform/adapter/in/web/BoardController.java
@@ -3,9 +3,9 @@ package com.zip.community.platform.adapter.in.web;
 import com.zip.community.common.response.ApiResponse;
 import com.zip.community.common.response.pageable.PageRequest;
 import com.zip.community.common.response.pageable.PageResponse;
-import com.zip.community.platform.adapter.in.web.dto.response.BoardListResponse;
+import com.zip.community.platform.adapter.in.web.dto.response.board.BoardListResponse;
 import com.zip.community.platform.adapter.in.web.dto.request.board.BoardRequest;
-import com.zip.community.platform.adapter.in.web.dto.response.BoardDetailResponse;
+import com.zip.community.platform.adapter.in.web.dto.response.board.BoardDetailResponse;
 import com.zip.community.platform.application.port.in.board.CreateBoardUseCase;
 import com.zip.community.platform.application.port.in.board.GetBoardUseCase;
 import com.zip.community.platform.application.port.in.board.RemoveBoardUseCase;
