+++ src/main/java/com/zip/community/platform/adapter/in/web/BoardTempController.java
@@ -4,8 +4,8 @@ import com.zip.community.common.response.ApiResponse;
 import com.zip.community.common.response.CustomException;
 import com.zip.community.common.response.errorcode.BoardErrorCode;
 import com.zip.community.platform.adapter.in.web.dto.request.board.TempBoardRequest;
-import com.zip.community.platform.adapter.in.web.dto.response.TempBoardDetailResponse;
-import com.zip.community.platform.adapter.in.web.dto.response.TempBoardListResponse;
+import com.zip.community.platform.adapter.in.web.dto.response.board.TempBoardDetailResponse;
+import com.zip.community.platform.adapter.in.web.dto.response.board.TempBoardListResponse;
 import com.zip.community.platform.application.port.in.board.TempBoardUseCase;
 import com.zip.community.platform.domain.board.Board;
 import lombok.RequiredArgsConstructor;
