+++ src/main/java/com/zip/community/platform/adapter/in/web/dto/response/board/CategoryResponse.java
@@ -1,4 +1,4 @@
-package com.zip.community.platform.adapter.in.web.dto.response;
+package com.zip.community.platform.adapter.in.web.dto.response.board;
 
 import com.zip.community.platform.domain.board.Category;
 import lombok.AllArgsConstructor;
