+++ src/main/java/com/zip/community/platform/adapter/in/web/dto/response/board/TempBoardDetailResponse.java
@@ -1,4 +1,4 @@
-package com.zip.community.platform.adapter.in.web.dto.response;
+package com.zip.community.platform.adapter.in.web.dto.response.board;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.zip.community.platform.domain.board.Board;
@@ -11,20 +11,22 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class TempBoardListResponse {
+public class TempBoardDetailResponse {
 
-    private Long category; // Immutable 빈 리스트 사용
+    private Long category;
 
     private Long author;
+
     private String title;
+    private String content;
 
     @Builder.Default
     private String thumbnailUrl = "";
-
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime createdAt;
@@ -34,22 +36,24 @@ public class TempBoardListResponse {
 
 
     // 생성자
-    public static TempBoardListResponse from(Board board) {
+    public static TempBoardDetailResponse from(Board board) {
 
-        return TempBoardListResponse.builder()
+        return TempBoardDetailResponse.builder()
                 .category(board.getCategoryId())
                 .title(board.getSnippet().getTitle())
+                .content(board.getSnippet().getContent())
                 .author(board.getMemberId())
                 .createdAt(board.getCreatedAt())
                 .updatedAt(board.getUpdatedAt())
                 .build();
     }
 
-    // List 해결
-    public static List<TempBoardListResponse> from(List<Board> boards) {
+    // 리스트 변환
+    public static List<TempBoardDetailResponse> from(List<Board> boards) {
         return boards.stream()
-                .map(TempBoardListResponse::from)
+                .map(TempBoardDetailResponse::from)
                 .collect(Collectors.toList());
     }
+
 
 }
